Rename relaying service and constants
